MVector
MVector is a custom Vector struct for the MW Unity Namespace to reduce the unnecessary functions of UnityEngine.Vector3 whilst providing some useful albeit specific new functionalities.

These new functionalities include the ability to directly ignore a component of a MVector. With UnityEngine.Vector3, it is sometimes useful to completely ignore the Z-Axis when computing something; with MVector.XY, among others, this is now possible with ease.

Another new functionality is the simple invocation of the cross and dot product of a vector by overloading the operator ‘^’ and ‘|’, respectively.

Others include checking whether an MVector is normalised and computing the UnityEngine.Quaternion equivalent that represents an MVector.

It should be noted that MW Unity Namespace does not intend to reinvent the wheel – there exists methods to convert MVectors to UnityEngine.Vector3 and vice versa. There may not be a required function that is built into MVector, but is in UnityEngine.Vector3. MVector is only designed to reduce the typing of long but simple computations that should be otherwise shortened.