Custom Player Behaviour
Player is a new MonoBehaviour that provides the base properties of what a player might have, including Movement and Health.

MPlayer inherits Player, providing more base functionality for a 3D game environment with default forward/right movement implementations with Rigidbody.

MPlayer2D is a variant of MPlayer. Is also inherits Player, but has slightly different default implementations to MPlayer to account for 2D movement. It also has implementations for flipping a SpriteRenderer over the X axis, based on input.