MW Modifications
A number of modifications were made to the structure of the MW Unity Namespace.

One of which was removing the broad MW tag. Everything that was in MW has now been moved to the MW.Core namespace, or another namespace.

A number of previously un-XML-commented areas of the MW namespace have now been commented.