Operator Overload and Decoration Support
Added support for overloaded operators to be read and shown in the Documentation.

Added support for 'Decorations' to classes, functions, etc. to be shown in the Documentation. Currently MVector is the only class with these Decorations.

Also changed the spacing between headings and their underneath summaries. They are now closer together, and further apart from the heading above/below it, and generally more readable.