MW Build Initialisation and Automation
If MW is being cloned and built for the first time, there is now a dedicated Initialise batch script that assists in doing so.

This should warn of any prerequisites and missing binaries or outputs. It also builds the required MW-dependent projects, such as MGenerator, MTest, and MWEditor. to ensure that building MW is faultless and, in the case of an error, disambiguous.